Full API

Autogenerated API list

Gabs.GaussianChannel — Type

Defines a Gaussian channel for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

basis: Symplectic representation for Gaussian channel.

disp: The displacement vector of length 2N.

transform: The transformation matrix of size 2N x 2N.

noise: The noise matrix of size 2N x 2N.

Mathematical description of a Gaussian channel

An $N$-mode Gaussian channel, $G(\mathbf{d}, \mathbf{T}, \mathbf{N})$, is an operator characterized by a displacement vector $\mathbf{d}$ of length $2N$, as well as a transformation matrix $\mathbf{T}$ and noise matrix $\mathbf{N}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian channel action on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{T} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{T} \mathbf{V} \mathbf{T}^{\text{T}} + \mathbf{N}.\]

Example

julia> noise = [1.0 -3.0; 4.0 2.0];

julia> displace(QuadPairBasis(1), 1.0+im, noise)
GaussianChannel for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
transform: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0
noise: 2×2 Matrix{Float64}:
 1.0  -3.0
 4.0   2.0

Gabs.GaussianState — Type

Defines a Gaussian state for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

basis: Symplectic basis for Gaussian state.

mean: The mean vector of length 2N.

covar: The covariance matrix of size 2N x 2N.

Mathematical description of a Gaussian state

An $N$-mode Gaussian state, $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$, is a density operator characterized by two statistical moments: a mean vector $\mathbf{\bar{x}}$ of length $2N$ and covariance matrix $\mathbf{V}$ of size $2N\times 2N$. By definition, the Wigner representation of a Gaussian state is a Gaussian function.

Example

julia> coherentstate(QuadPairBasis(1), 1.0+im)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.GaussianUnitary — Type

Defines a Gaussian unitary for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

basis: Symplectic basis for Gaussian unitary.

disp: The displacement vector of length 2N.

symplectic: The symplectic matrix of size 2N x 2N.

Mathematical description of a Gaussian unitary

An $N$-mode Gaussian unitary, $U(\mathbf{d}, \mathbf{S})$, is a unitary operator characterized by a displacement vector $\mathbf{d}$ of length $2N$ and symplectic matrix $\mathbf{S}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian unitary transformation on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{S} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{S} \mathbf{V} \mathbf{S}^{\text{T}}.\]

Example

julia> displace(QuadPairBasis(1), 1.0+im)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.Generaldyne — Type

Defines a partial general-dyne measurement of a Gaussian system.

Fields

system: Initial Gaussian state.

outcome: Measurement outcome Gaussian state.

indices: Specific indices of system that are projected onto outcome.

Mathematical description of a general-dyne measurement

A partial general-dyne measurement on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ of a system partitioned into subsystems $A$ and $B$ is a projection of subsystem $B$ onto a measurement outcome Gaussian state $\hat{\rho}(\mathbf{\bar{x}}_{m}, \mathbf{V}_{m})$.

Example

julia> basis = QuadPairBasis(1);

julia> vac = vacuumstate(basis); coh = coherentstate(basis, 1.0-im);

julia> state = vac ⊗ coh ⊗ vac ⊗ coh
GaussianState for 4 modes.
  symplectic basis: QuadPairBasis
mean: 8-element Vector{Float64}:
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
covariance: 8×8 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0

julia> Generaldyne(state, coh ⊗ vac ⊗ coh, [1, 3, 4])
Generaldyne on indices [1, 3, 4]
system: GaussianState for 4 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})
outcome: GaussianState for 3 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})

Gabs.QuadBlockBasis — Type

Defines a symplectic basis for a bosonic system of size nmodes in which the quadrature field operators are arranged blockwise.

Gabs.QuadPairBasis — Type

Defines a symplectic basis for a bosonic system of size nmodes in which the quadrature field operators are arranged pairwise.

Gabs.amplifier — Method

amplifier([Td=Vector{Float64}, Tt=Matrix{Float64},] basis::SymplecticBasis, r<:Real, n<:Int)

Gaussian channel describing the interaction of an input single mode Gaussian state and its environment via a two-mode squeezing operation. The channel is paramatrized by squeezing amplitude parameter r and thermal noise n.

Mathematical description of an amplifier channel

An amplifier channel, $\mathcal{A}_{\theta}^{n_{\text{th}}}$, where $r$ is the squeezing amplitude parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cosh r\mathbf{I}, \qquad \mathbf{N} = (\sinh r)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> amplifier(QuadPairBasis(1), 2.0, 3)
GaussianChannel for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 3.7622  0.0
 0.0     3.7622
noise: 2×2 Matrix{Float64}:
 39.4623   0.0
  0.0     39.4623

Gabs.attenuator — Method

attenuator([Td=Vector{Float64}, Tt=Matrix{Float64},] basis::SymplecticBasis, theta<:Real, n<:Int)

Gaussian channel describing the coupling of an input single mode Gaussian state and its environment via a beam splitter operation. The channel is paramatrized by beam splitter rotation angle theta and thermal noise n.

Mathematical description of an attenuator channel

An attenuator channel, $\mathcal{E}_{\theta}^{n_{\text{th}}}$, where $\theta$ is the beam splitter rotation parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cos\theta\mathbf{I}, \qquad \mathbf{N} = (\sin\theta)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> attenuator(QuadPairBasis(1), pi/6, 3)
GaussianChannel for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 0.866025  0.0
 0.0       0.866025
noise: 2×2 Matrix{Float64}:
 0.75  0.0
 0.0   0.75

Gabs.beamsplitter — Method

beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, transmit<:Real)
beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, transmit<:Real, noise::Ts)

Gaussian operator that serves as the beam splitter transformation of a two-mode Gaussian state, known as the beam splitter operator. The symplectic representation is given by basis. The transmittivity of the operator is given by transmit. Noise can be added to the operation with noise.

Mathematical description of a beam splitter operator

A beam splitter operator $B(\tau)$ is defined by the operation $B(\theta) = \exp\left[\theta(\hat{a}^{\dagger}\hat{b} - \hat{a}\hat{b}^{\dagger})\right]$, where $\theta$ is defined by $\tau = \cos^2\theta$, and $\hat{a}$ and $\hat{b}$ are the annihilation operators of the two modes, respectively. The operator $B(\tau)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \sqrt{\tau}\mathbf{I} & \sqrt{1-\tau}\mathbf{I} \\ -\sqrt{1-\tau}\mathbf{I} & \sqrt{\tau}\mathbf{I} \\ \end{pmatrix}.\]

Example

julia> beamsplitter(QuadPairBasis(2), 0.75)
GaussianUnitary for 2 modes.
  symplectic basis: QuadPairBasis
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  0.5        0.0       0.866025  0.0
  0.0        0.5       0.0       0.866025
 -0.866025   0.0       0.5       0.0
  0.0       -0.866025  0.0       0.5

Gabs.coherentstate — Method

coherentstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, alpha<:Number)

Gaussian state that is the quantum analogue of a monochromatic electromagnetic field, known as the coherent state. The symplectic representation is defined by basis. The complex amplitude of the state is given by alpha.

Mathematical description of a coherent state

A coherent state $|\alpha\rangle$, where $\alpha$ is the complex amplitude, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> coherentstate(QuadPairBasis(1), 1.0+im)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.displace — Method

displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, alpha<:Number)
displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, alpha<:Number, noise::Ts)

Gaussian operator that displaces the vacuum state into a coherent state, known as the displacement operator. The symplectic representation is given by basis. The complex amplitude is given by alpha. Noise can be added to the operation with noise.

Mathematical description of a displacement operator

A displacement operator $D(\alpha)$ is defined by the operation $D(\alpha)|0\rangle = |\alpha\rangle$, where $\alpha$ is a complex amplitude. The operator $D(\alpha)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{S} = \mathbf{I}.\]

Example

julia> displace(QuadPairBasis(1), 1.0+im)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.eprstate — Method

eprstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, r<:Real, theta<:Real)

Gaussian state that is a two-mode squeezed state, known as the Einstein-Podolski-Rosen (EPR) state. The symplectic representation is defined by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of an EPR state

An EPR state $|r, \theta\rangle_{\text{EPR}}$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\begin{pmatrix} \cosh(2r)\mathbf{I} & -\sinh(2r)\mathbf{R}(\theta) \\ -\sinh(2r)\mathbf{R}(\theta) & \cosh(2r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> eprstate(QuadPairBasis(2), 0.5, pi/4)
GaussianState for 2 modes.
  symplectic basis: QuadPairBasis
mean: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
  0.77154    0.0       -0.415496  -0.415496
  0.0        0.77154   -0.415496   0.415496
 -0.415496  -0.415496   0.77154    0.0
 -0.415496   0.415496   0.0        0.77154

Gabs.isgaussian — Method

isgaussian(x::GaussianState)
isgaussian(x::GaussianUnitary)
isgaussian(x::GaussianChannel)

Check if x satisfies the corresponding Gaussian definition for its type.

Example

julia> basis = QuadPairBasis(1);

julia> op = displace(basis, 1.0-im)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
  1.4142135623730951
 -1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

julia> isgaussian(op)
true

Gabs.issymplectic — Method

issymplectic(basis::SymplecticBasis, x::T)

Check if input matrix satisfies symplectic definition.

Example

julia> basis = QuadPairBasis(1);

julia> issymplectic(basis, [1.0 0.0; 0.0 1.0])
true

Gabs.output — Method

output(meas::Generaldyne)

Conditional mapping of the subsystem for the initial Gaussian system that is not measured in a Generaldyne object.

Mathematical description of conditional evolution of a Gaussian system

Let $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ be a Gaussian state of a system partitioned into subsystems $A$ and $B$, such that

\[\mathbf{\bar{x}} = \begin{pmatrix} \mathbf{\bar{x}}_{\text{A}} & \mathbf{\bar{x}}_{\text{A}} \end{pmatrix}^{\text{T}}, \quad \mathbf{V} = \begin{pmatrix} \mathbf{V}_{\text{A}} & \mathbf{V}_{\text{AB}} \\ \mathbf{V}_{\text{AB}}^{\text{T}} & \mathbf{V}_{\text{B}} \end{pmatrix},\]

where $(\mathbf{\bar{x}}_{\text{A}}, \mathbf{V}_{\text{A}})$ and $(\mathbf{\bar{x}}_{\text{B}}, \mathbf{V}_{\text{B}})$ are the mean and covariance matrices of subsystems $A$ and $B$, respectively. Here $\mathbf{V}_{\text{AB}}$ is a covariance matrix representing the correlations between $A$ and $B$. If we project subsystem $B$ onto a measurement outcome $\hat{\rho}(\mathbf{\bar{x}}_{m}, \mathbf{V}_{m})$, then subsystem $A$ conditionally maps as follows:

\[\mathbf{\bar{x}}_{\text{A}} \to \mathbf{\bar{x}}_{\text{A}} + \frac{\mathbf{V}_{\text{AB}}}{\mathbf{V}_{\text{B}} + \mathbf{V}_{m}}(\mathbf{\bar{x}}_{m} - \mathbf{\bar{x}}_{\text{B}})^{\text{T}}, \quad\mathbf{V}_{\text{A}} \to \mathbf{V}_{\text{A}} - \frac{\mathbf{V}_{\text{AB}}}{\mathbf{V}_{\text{B}} + \mathbf{V}_{m}} \mathbf{V}_{\text{AB}}^{\text{T}}.\]

Example

julia> basis = QuadPairBasis(1);

julia> vac = vacuumstate(basis); coh = coherentstate(basis, 1.0-im);

julia> state = vac ⊗ coh ⊗ vac ⊗ coh
GaussianState for 4 modes.
  symplectic basis: QuadPairBasis
mean: 8-element Vector{Float64}:
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
covariance: 8×8 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0

julia> gd = Generaldyne(state, coh ⊗ vac ⊗ coh, [1, 3, 4])
Generaldyne on indices [1, 3, 4]
system: GaussianState for 4 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})
outcome: GaussianState for 3 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})

julia> output(gd)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
  1.4142135623730951
 -1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.phaseshift — Method

phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, theta<:Real)
phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, theta<:Real, noise::Ts)

Gaussian operator that rotates the phase of a given Gaussian mode by theta, as the phase shift operator. The symplectic representation is given by basis. Noise can be added to the operation with noise.

Mathematical description of a phase shift operator

A phase shift operator $U(\theta)$ is defined by the operation $U(\Theta) = \exp(-i\theta\hat{a}^{\dagger}\hat{a})$, where $\theta$ is the phase parameter, and $\hat{a}^{\dagger}$ and $\hat{a}$ are the raising and lowering operators, respectively. The operator $U(\theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cos(\theta) & \sin(\theta) \\ -\sin(\theta) & \cos(\theta) \\ \end{pmatrix}.\]

Example

julia> phaseshift(QuadPairBasis(1), 3pi/4)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
 -0.707107   0.707107
 -0.707107  -0.707107

Gabs.prob — Method

prob(meas::Generaldyne)

Calculate the probability density of a general-dyne measurement.

Mathematical description of probability density of a general-dyne measurement

Let $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ be a Gaussian state of a system partitioned into subsystems $A$ and $B$, such that

\[\mathbf{\bar{x}} = \begin{pmatrix} \mathbf{\bar{x}}_{\text{A}} & \mathbf{\bar{x}}_{\text{A}} \end{pmatrix}^{\text{T}}, \quad \mathbf{V} = \begin{pmatrix} \mathbf{V}_{\text{A}} & \mathbf{V}_{\text{AB}} \\ \mathbf{V}_{\text{AB}}^{\text{T}} & \mathbf{V}_{\text{B}} \end{pmatrix},\]

where $(\mathbf{\bar{x}}_{\text{A}}, \mathbf{V}_{\text{A}})$ and $(\mathbf{\bar{x}}_{\text{B}}, \mathbf{V}_{\text{B}})$ are the mean and covariance matrices of subsystems $A$ and $B$, respectively. Here $\mathbf{V}_{\text{AB}}$ is a covariance matrix representing the correlations between $A$ and $B$. If we project subsystem $B$ onto a measurement outcome $\hat{\rho}(\mathbf{\bar{x}}_{m}, \mathbf{V}_{m})$, then the probability density in $d\mathbf{\hat{x}}_{m}$ is

\[p(\mathbf{\bar{x}}_{m}) = \frac{e^{(\mathbf{\bar{x}}_{m} - \mathbf{\bar{x}}_{\text{B}})^{\text{T}} \frac{1}{\mathbf{V}_{\text{B}} + \mathbf{V}_{m}} (\mathbf{\bar{x}}_{m} - \mathbf{\bar{x}}_{\text{B}})}}{\pi^{m} \sqrt{\det(\mathbf{V}_{\text{B}} + \mathbf{V}_{m})}},\]

where $m$ is the number of modes in subsystem $B$.

Example

julia> basis = QuadPairBasis(1);

julia> vac = vacuumstate(basis); coh = coherentstate(basis, 1.0-im);

julia> state = vac ⊗ coh ⊗ vac ⊗ coh
GaussianState for 4 modes.
  symplectic basis: QuadPairBasis
mean: 8-element Vector{Float64}:
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
  0.0
  0.0
  1.4142135623730951
 -1.4142135623730951
covariance: 8×8 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  1.0

julia> gd = Generaldyne(state, coh ⊗ vac ⊗ coh, [1, 3, 4])
Generaldyne on indices [1, 3, 4]
system: GaussianState for 4 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})
outcome: GaussianState for 3 modes.
 (xType: Vector{Float64} | vType: Matrix{Float64})

julia> prob(gd)
0.029788549650438086

Gabs.ptrace — Method

ptrace([Tm=Vector{Float64}, Tc=Matrix{Float64},] state::GaussianState, idx<:Int)
ptrace([Tm=Vector{Float64}, Tc=Matrix{Float64},] state::GaussianState, indices<:AbstractVector)

Partial trace of a Gaussian state over a subsystem indicated by idx, or multiple subsystems indicated by indices.

Example

julia> basis = QuadPairBasis(1);

julia> state = coherentstate(basis, 1.0+im) ⊗ thermalstate(basis, 2) ⊗ squeezedstate(basis, 3.0, pi/4)
GaussianState for 3 modes.
  symplectic basis: QuadPairBasis
mean: 6-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
 0.0
 0.0
covariance: 6×6 Matrix{Float64}:
 1.0  0.0  0.0  0.0   0.0       0.0
 0.0  1.0  0.0  0.0   0.0       0.0
 0.0  0.0  2.5  0.0   0.0       0.0
 0.0  0.0  0.0  2.5   0.0       0.0
 0.0  0.0  0.0  0.0  29.5414   71.3164
 0.0  0.0  0.0  0.0  71.3164  172.174

julia> ptrace(state, 2)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 2.5  0.0
 0.0  2.5

julia> ptrace(state, [1, 3])
GaussianState for 2 modes.
  symplectic basis: QuadPairBasis
mean: 4-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
 1.0  0.0   0.0       0.0
 0.0  1.0   0.0       0.0
 0.0  0.0  29.5414   71.3164
 0.0  0.0  71.3164  172.174

Gabs.purity — Method

purity(state::GaussianState)

Calculate the purity of a Gaussian state, defined by 1/sqrt(det(V)).

Gabs.randchannel — Method

randchannel([Td=Vector{Float64}, Tt=Matrix{Float64},] basis::SymplecticBasis)

Calculate a random Gaussian channel in symplectic representation defined by basis.

Gabs.randstate — Method

randstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis; pure=false)

Calculate a random Gaussian state in symplectic representation defined by basis.

Gabs.randsymplectic — Method

randsymplectic([T=Matrix{Float64},] basis::SymplecticBasis, passive=false)

Calculate a random symplectic matrix in symplectic representation defined by basis.

Gabs.randunitary — Method

randunitary([Td=Vector{Float64}, Ts=Matrix{Float64},] basis::SymplecticBasis; passive=false)

Calculate a random Gaussian unitary operator in symplectic representation defined by basis.

Gabs.squeeze — Method

squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real)
squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes the vacuum state into a squeezed state, known as the squeezing operator. The symplectic representation is given by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a squeezing operator

A squeeze operator $S(r, \theta)$ is defined by the operation $S(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \cosh(r)\mathbf{I} - \sinh(r)\mathbf{R}(\theta),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeeze(QuadPairBasis(1), 0.25, pi/4)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
  0.852789  -0.178624
 -0.178624   1.21004

Gabs.squeezedstate — Method

squeezedstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, r<:Real, theta<:Real)

Gaussian state with quantum uncertainty in its phase and amplitude, known as the squeezed state. The symplectic representation is defined by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of a squeezed state

A squeezed state $|r, \theta\rangle$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\left(\cosh(2r)\mathbf{I} - \sinh(2r)\mathbf{R}(\theta)\right),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeezedstate(QuadPairBasis(1), 0.5, pi/4)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 0.356044  0.415496
 0.415496  1.18704

Gabs.symplecticform — Method

symplecticform([T = Matrix{Float64},] basis::SymplecticBasis)

Compute the symplectic form matrix of size 2N x 2N corresponding to basis.

Gabs.sympspectrum — Method

sympspectrum(state::GaussianState)

Compute the symplectic spectrum of a Gaussian state.

Gabs.thermalstate — Method

thermalstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, photons<:Int)

Gaussian state at thermal equilibrium, known as the thermal state. The symplectic representation is defined by basis. The mean photon number of the state is given by photons.

Mathematical description of a thermal state

A thermal state $|\bar{n}\rangle$, where $\bar{n}$ is the mean number of photons, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \left(\bar{n} + \frac{1}{2}\right)\mathbf{I}.\]

Example

julia> thermalstate(QuadPairBasis(1), 4)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 4.5  0.0
 0.0  4.5

Gabs.twosqueeze — Method

twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real)
twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes a two-mode vacuum state into a two-mode squeezed state, known as the two-mode squeezing operator. The symplectic representation is given by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a two-mode squeezing operator

A two-mode squeeze operator $S_2(r, \theta)$ is defined by the operation $S_2(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S_2(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cosh(r)\mathbf{I} & -\sinh(r)\mathbf{R}(\theta) \\ -\sinh(r)\mathbf{R}(\theta) & \cosh(r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> twosqueeze(QuadPairBasis(2), 0.25, pi/4)
GaussianUnitary for 2 modes.
  symplectic basis: QuadPairBasis
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  1.03141    0.0       -0.178624  -0.178624
  0.0        1.03141   -0.178624   0.178624
 -0.178624  -0.178624   1.03141    0.0
 -0.178624   0.178624   0.0        1.03141

Gabs.vacuumstate — Method

vacuumstate([Tm=Vector{Float64}, Tc=Matrix{Float64}], basis::SymplecticBasis)

Gaussian state with zero photons, known as the vacuum state. The symplectic representation is defined by basis.

Mathematical description of a vacuum state

A vacuum state $|0\rangle$ is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> vacuumstate(QuadPairBasis(1))
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.wigner — Method

wigner(state::GaussianState, x)

Compute the Wigner function of an N-mode Gaussian state at x, a vector of size 2N.

Gabs.wignerchar — Method

wignerchar(state::GaussianState, xi)

Compute the Wigner characteristic function of an N-mode Gaussian state at xi, a vector of size 2N.

QuantumInterface.directsum — Method

directsum(basis1::SymplecticBasis, basis2::SymplecticBasis)

Compute the direct sum of symplectic bases.

QuantumInterface.tensor — Method

tensor(state1::GaussianState, state2::GaussianState)

tensor product of Gaussian states, which can also be called with ⊗.

Example

julia> basis = QuadPairBasis(1);

julia> coherentstate(basis, 1.0+im) ⊗ thermalstate(basis, 2)
GaussianState for 2 modes.
  symplectic basis: QuadPairBasis
mean: 4-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  2.5  0.0
 0.0  0.0  0.0  2.5