Full API

Autogenerated API list

Gabs.GaussianChannel — Type

Defines a Gaussian channel for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

disp: The displacement vector of length 2N.

transform: The transformation matrix of size 2N x 2N.

noise: The noise matrix of size 2N x 2N.

nmodes: The number of modes N for the Gaussian channel.

Mathematical description of a Gaussian channel

An $N$-mode Gaussian channel, $G(\mathbf{d}, \mathbf{T}, \mathbf{N})$, is an operator characterized by a displacement vector $\mathbf{d}$ of length $2N$, as well as a transformation matrix $\mathbf{T}$ and noise matrix $\mathbf{N}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian channel action on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{S} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{T} \mathbf{V} \mathbf{T}^{\text{T}} + \mathbf{N}.\]

Example

julia> noise = [1.0 -3.0; 4.0 2.0];

julia> displace(1.0+im, noise)
GaussianChannel for 1 mode.
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
transform: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0
noise: 2×2 Matrix{Float64}:
 1.0  -3.0
 4.0   2.0

Gabs.GaussianState — Type

Defines a Gaussian state for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

mean: The mean vector of length 2N.

covar: The covariance matrix of size 2N x 2N.

nmodes: The number of modes N of the Gaussian state.

Mathematical description of a Gaussian state

An $N$-mode Gaussian state, $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$, is a density operator characterized by two statistical moments: a mean vector $\mathbf{\bar{x}}$ of length $2N$ and covariance matrix $\mathbf{V}$ of size $2N\times 2N$. By definition, the Wigner representation of a Gaussian state is a Gaussian function.

Example

julia> coherentstate(1.0+im)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.GaussianUnitary — Type

Defines a Gaussian unitary for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

disp: The displacement vector of length 2N.

symplectic: The symplectic matrix of size 2N x 2N.

nmodes: The number of modes N for the Gaussian unitary.

Mathematical description of a Gaussian unitary

An $N$-mode Gaussian unitary, $U(\mathbf{d}, \mathbf{S})$, is a unitary operator characterized by a displacement vector $\mathbf{d}$ of length $2N$ and symplectic matrix $\mathbf{S}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian unitary transformation on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{S} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{S} \mathbf{V} \mathbf{S}^{\text{T}}.\]

Example

julia> displace(1.0+im)
GaussianUnitary for 1 mode.
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.amplifier — Method

amplifier([Td=Vector{Float64}, Tt=Matrix{Float64},] r<:Real, n<:Int)

Gaussian channel describing the interaction of an input single mode Gaussian state and its environment via a two-mode squeezing operation. The channel is paramatrized by squeezing amplitude parameter r and thermal noise n.

Mathematical description of an amplifier channel

An amplifier channel, $\mathcal{A}_{\theta}^{n_{\text{th}}}$, where $r$ is the squeezing amplitude parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cosh r\mathbf{I}, \qquad \mathbf{N} = (\sinh r)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> amplifier(2.0, 3)
GaussianChannel for 1 mode.
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 3.7622  0.0
 0.0     3.7622
noise: 2×2 Matrix{Float64}:
 39.4623   0.0
  0.0     39.4623

Gabs.attenuator — Method

attenuator([Td=Vector{Float64}, Tt=Matrix{Float64},] theta<:Real, n<:Int)

Gaussian channel describing the coupling of an input single mode Gaussian state and its environment via a beam splitter operation. The channel is paramatrized by beam splitter rotation angle theta and thermal noise n.

Mathematical description of an attenuator channel

An attenuator channel, $\mathcal{E}_{\theta}^{n_{\text{th}}}$, where $\theta$ is the beam splitter rotation parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cos\theta\mathbf{I}, \qquad \mathbf{N} = (\sin\theta)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> attenuator(pi/6, 3)
GaussianChannel for 1 mode.
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 0.866025  0.0
 0.0       0.866025
noise: 2×2 Matrix{Float64}:
 0.75  0.0
 0.0   0.75

Gabs.beamsplitter — Method

beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], transmit<:Real)
beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], transmit<:Real, noise::Ts)

Gaussian operator that serves as the beam splitter transformation of a two-mode Gaussian state, known as the beam splitter operator. The transmittivity of the operator is given by transmit. Noise can be added to the operation with noise.

Mathematical description of a beam splitter operator

A beam splitter operator $B(\tau)$ is defined by the operation $B(\theta) = \exp\left[\theta(\hat{a}^{\dagger}\hat{b} - \hat{a}\hat{b}^{\dagger})\right]$, where $\theta$ is defined by $\tau = \cos^2\theta$, and $\hat{a}$ and $\hat{b}$ are the annihilation operators of the two modes, respectively. The operator $B(\tau)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \sqrt{\tau}\mathbf{I} & \sqrt{1-\tau}\mathbf{I} \\ -\sqrt{1-\tau}\mathbf{I} & \sqrt{\tau}\mathbf{I} \\ \end{pmatrix}.\]

Example

julia> beamsplitter(0.75)
GaussianUnitary for 2 modes.
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  0.866025   0.0       0.5       0.0
  0.0        0.866025  0.0       0.5
 -0.5       -0.0       0.866025  0.0
 -0.0       -0.5       0.0       0.866025

Gabs.coherentstate — Method

coherentstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] alpha<:Number)

Gaussian state that is the quantum analogue of a monochromatic electromagnetic field, known as the coherent state. The complex amplitude of the state is given by alpha.

Mathematical description of a coherent state

A coherent state $|\alpha\rangle$, where $\alpha$ is the complex amplitude, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> coherentstate(1.0+im)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.displace — Method

displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], alpha<:Number)
displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], alpha<:Number, noise::Ts)

Gaussian operator that displaces the vacuum state into a coherent state, known as the displacement operator. The complex amplitude is given by alpha. Noise can be added to the operation with noise.

Mathematical description of a displacement operator

A displacement operator $D(\alpha)$ is defined by the operation $D(\alpha)|0\rangle = |\alpha\rangle$, where $\alpha$ is a complex amplitude. The operator $D(\alpha)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{S} = \mathbf{I}.\]

Example

julia> displace(1.0+im)
GaussianUnitary for 1 mode.
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.eprstate — Method

eprstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] r<:Real, theta<:Real)

Gaussian state that is a two-mode squeezed state, known as the Einstein-Podolski-Rosen (EPR) state. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of an EPR state

An EPR state $|r, \theta\rangle_{\text{EPR}}$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\begin{pmatrix} \cosh(2r)\mathbf{I} & -\sinh(2r)\mathbf{R}(\theta) \\ -\sinh(2r)\mathbf{R}(\theta) & \cosh(2r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> eprstate(0.5, pi/4)
GaussianState for 2 modes.
mean: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
 0.77154    0.0       0.415496   0.415496
 0.0        0.77154   0.415496  -0.415496
 0.415496   0.415496  0.77154    0.0
 0.415496  -0.415496  0.0        0.77154

Gabs.phaseshift — Method

phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], theta<:Real)
phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], theta<:Real, noise::Ts)

Gaussian operator that rotates the phase of a given Gaussian mode by theta, as the phase shift operator. Noise can be added to the operation with noise.

Mathematical description of a phase shift operator

A phase shift operator $U(\theta)$ is defined by the operation $U(\Theta) = \exp(-i\theta\hat{a}^{\dagger}\hat{a})$, where $\theta$ is the phase parameter, and $\hat{a}^{\dagger}$ and $\hat{a}$ are the raising and lowering operators, respectively. The operator $U(\theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cos(\theta) & \sin(\theta) \\ -\sin(\theta) & \cos(\theta) \\ \end{pmatrix}.\]

Example

julia> phaseshift(3pi/4)
GaussianUnitary for 1 mode.
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
 -0.707107   0.707107
 -0.707107  -0.707107

Gabs.ptrace — Method

ptrace([Tm=Vector{Float64}, Tc=Matrix{Float64},] state::GaussianState, idx<:Int)
ptrace([Tm=Vector{Float64}, Tc=Matrix{Float64},] state::GaussianState, indices<:AbstractVector)

Partial trace of a Gaussian state over a subsytem indicated by idx, or multiple subsystems indicated by indices.

Example

julia> state = coherentstate(1.0+im) ⊗ thermalstate(2) ⊗ squeezedstate(3.0, pi/4)
GaussianState for 3 modes.
mean: 6-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
 0.0
 0.0
covariance: 6×6 Matrix{Float64}:
 1.0  0.0  0.0  0.0   0.0       0.0
 0.0  1.0  0.0  0.0   0.0       0.0
 0.0  0.0  2.5  0.0   0.0       0.0
 0.0  0.0  0.0  2.5   0.0       0.0
 0.0  0.0  0.0  0.0  29.5414   71.3164
 0.0  0.0  0.0  0.0  71.3164  172.174

julia> ptrace(state, 2)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 2.5  0.0
 0.0  2.5

julia> ptrace(state, [1, 3])
GaussianState for 2 modes.
mean: 4-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
 1.0  0.0   0.0       0.0
 0.0  1.0   0.0       0.0
 0.0  0.0  29.5414   71.3164
 0.0  0.0  71.3164  172.174

Gabs.squeeze — Method

squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], r<:Real, theta<:Real)
squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes the vacuum state into a squeezed state, known as the squeezing operator. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a squeezing operator

A squeeze operator $S(r, \theta)$ is defined by the operation $S(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \cosh(r)\mathbf{I} - \sinh(r)\mathbf{R}(\theta),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeeze(0.25, pi/4)
GaussianUnitary for 1 mode.
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
 0.215425   0.0451226
 0.0451226  0.30567

Gabs.squeezedstate — Method

squeezedstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] r<:Real, theta<:Real)

Gaussian state with quantum uncertainty in its phase and amplitude, known as the squeezed state. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of a squeezed state

A squeezed state $|r, \theta\rangle$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\left(\cosh(2r)\mathbf{I} - \sinh(2r)\mathbf{R}(\theta)\right),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeezedstate(0.5, pi/4)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 0.356044  0.415496
 0.415496  1.18704

Gabs.symplecticform — Method

symplecticform([T = Matrix{Float64},] modes<:Int)

Compute the symplectic form matrix of size 2N x 2N, where N is given by modes.

Gabs.thermalstate — Method

thermalstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] photons<:Int)

Gaussian state at thermal equilibrium, known as the thermal state. The mean photon number of the state is given by photons.

Mathematical description of a thermal state

A thermal state $|\bar{n}\rangle$, where $\bar{n}$ is the mean number of photons, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \left(\bar{n} + \frac{1}{2}\right)\mathbf{I}.\]

Example

julia> thermalstate(4)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 4.5  0.0
 0.0  4.5

Gabs.twosqueeze — Method

twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], r<:Real, theta<:Real)
twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes a two-mode vacuum state into a two-mode squeezed state, known as the two-mode squeezing operator. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a two-mode squeezing operator

A two-mode squeeze operator $S_2(r, \theta)$ is defined by the operation $S_2(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S_2(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cosh(r)\mathbf{I} & -\sinh(r)\mathbf{R}(\theta) \\ -\sinh(r)\mathbf{R}(\theta) & \cosh(r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> twosqueeze(0.25, pi/4)
GaussianUnitary for 2 modes.
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  1.03141    0.0       -0.178624  -0.178624
  0.0        1.03141   -0.178624   0.178624
 -0.178624  -0.178624   1.03141    0.0
 -0.178624   0.178624   0.0        1.03141

Gabs.vacuumstate — Method

vacuumstate([Tm=Vector{Float64}, Tc=Matrix{Float64}])

Gaussian state with zero photons, known as the vacuum state.

Mathematical description of a vacuum state

A vacuum state $|0\rangle$ is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> vacuumstate()
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.wigner — Method

wigner(state::GaussianState, x)

Compute the Wigner function of an N-mode Gaussian state at x, a vector of size 2N.

Gabs.wignerchar — Method

wignerchar(state::GaussianState, xi)

Compute the Wigner characteristic function of an N-mode Gaussian state at xi, a vector of size 2N.

QuantumInterface.tensor — Method

tensor(state1::GaussianState, state2::GaussianState)

tensor product of Gaussian states, which can also be called with ⊗.

Example

julia> coherentstate(1.0+im) ⊗ thermalstate(2)
GaussianState for 2 modes.
mean: 4-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  2.5  0.0
 0.0  0.0  0.0  2.5