Manual

Simply put, Gabs.jl is a package for creating and transforming Gaussian bosonic systems. This section discusses the "lower level" tools for simulating such phenomena, with mathematical explanations when appropriate. For comprehensive reviews of Gaussian quantum information, see the suggested readings page.

Gaussian States

The star of this package is the GaussianState type, which allows us to initialize and manipulate a phase space description of an arbitrary Gaussian state.

Gabs.GaussianState — Type

Defines a Gaussian state for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

mean: The mean vector of length 2N.

covar: The covariance matrix of size 2N x 2N.

nmodes: The number of modes N of the Gaussian state.

Mathematical description of a Gaussian state

An $N$-mode Gaussian state, $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$, is a density operator characterized by two statistical moments: a mean vector $\mathbf{\bar{x}}$ of length $2N$ and covariance matrix $\mathbf{V}$ of size $2N\times 2N$. By definition, the Wigner representation of a Gaussian state is a Gaussian function.

Example

julia> coherentstate(1.0+im)
GaussianState for 1 mode.
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Functions to create instances of elementary Gaussian states are provided as part of the package API. Listed below are supported predefined Gaussian states:

vacuumstate

thermalstate

coherentstate

squeezedstate

eprstate

Detailed discussions and mathematical descriptions for each of these states are given in the Gaussian Zoos page.

If we were operating in the state (Fock) space, and wanted to describe multi-mode Gaussian states, we would take the tensor product of multiple density operators. That method, however, is quite computationally expensive and requires a finite truncation of the Fock basis. To create such state vector simulations, we recommend using the QuantumOptics.jl library. For our purposes in the phase space, we can manually create multi-mode Gaussian systems with a tensor product, which can be called with either tensor or ⊗, the Kronecker product symbol which can be typed in the Julia REPL as \otimes<TAB>. Take the following example, where we produce a 3-mode Gaussian state that consists of a coherent state, vacuumstate, and squeezed state:

julia> coherentstate(-1.0) ⊗ vacuumstate() ⊗ squeezedstate(0.25, pi/4)
GaussianState for 3 modes.
mean: 6-element Vector{Float64}:
 -1.4142135623730951
  0.0
  0.0
  0.0
  0.0
  0.0
covariance: 6×6 Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0       0.0
 0.0  1.0  0.0  0.0  0.0       0.0
 0.0  0.0  1.0  0.0  0.0       0.0
 0.0  0.0  0.0  1.0  0.0       0.0
 0.0  0.0  0.0  0.0  0.379578  0.184235
 0.0  0.0  0.0  0.0  0.184235  0.748048

Gaussian Operators

To transform Gaussian states into Gaussian states, we need Gaussian maps. There are various ways to construct Gaussian transformations, which we will discuss in this section.

Gaussian Unitaries

Let's begin with the simplest Gaussian transformation, a unitary transformation, which can be created with the GaussianUnitary type:

Gabs.GaussianUnitary — Type

Defines a Gaussian unitary for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

disp: The displacement vector of length 2N.

symplectic: The symplectic matrix of size 2N x 2N.

nmodes: The number of modes N for the Gaussian unitary.

Mathematical description of a Gaussian unitary

An $N$-mode Gaussian unitary, $U(\mathbf{d}, \mathbf{S})$, is a unitary operator characterized by a displacement vector $\mathbf{d}$ of length $2N$ and symplectic matrix $\mathbf{S}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian unitary transformation on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{S} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{S} \mathbf{V} \mathbf{S}^{\text{T}}.\]

Example

julia> displace(1.0+im)
GaussianUnitary for 1 mode.
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

This is a rather clean way to characterize a large group of Gaussian transformations on an N-mode Gaussian bosonic system. As long as we have a displacement vector of size 2N and symplectic matrix of size 2N x 2N, we can create a Gaussian transformation.

Note

A matrix $\mathbf{S}$ of size $2N\times 2N$ is symplectic when it satisfies the following relation:

\[\mathbf{S} \mathbf{\Omega} \mathbf{S}^{\text{T}} = \mathbf{\Omega}, \qquad \mathbf{\Omega} \equiv \bigoplus_{i=1}^{N} \begin{pmatrix} 0 & -1 \\ -1 & 0 \end{pmatrix}.\]

This library has a number of predefined Gaussian unitaries, which are listed below:

displace

squeeze

twosqueeze

phaseshift

beamsplitter

Detailed discussions and mathematical descriptions for each of these unitaries are given in the Gaussian Zoos page.

Gaussian Channels

Noisy bosonic channels are an important model for describing the interaction between a Gaussian state and its environment. Similar to Gaussian unitaries, Gaussian channels are linear bosonic channels that map Gaussian states to Gaussian states. Such objects can be created with the GaussianChannel type:

Gabs.GaussianChannel — Type

Defines a Gaussian channel for an N-mode bosonic system over a 2N-dimensional phase space.

Fields

disp: The displacement vector of length 2N.

transform: The transformation matrix of size 2N x 2N.

noise: The noise matrix of size 2N x 2N.

nmodes: The number of modes N for the Gaussian channel.

Mathematical description of a Gaussian channel

An $N$-mode Gaussian channel, $G(\mathbf{d}, \mathbf{T}, \mathbf{N})$, is an operator characterized by a displacement vector $\mathbf{d}$ of length $2N$, as well as a transformation matrix $\mathbf{T}$ and noise matrix $\mathbf{N}$ of size $2N\times 2N$, such that its action on a Gaussian state results in a Gaussian state. More specifically, a Gaussian channel action on a Gaussian state $\hat{\rho}(\mathbf{\bar{x}}, \mathbf{V})$ is described by its maps on the statistical moments of the Gaussian state:

\[\mathbf{\bar{x}} \to \mathbf{S} \mathbf{\bar{x}} + \mathbf{d}, \quad \mathbf{V} \to \mathbf{T} \mathbf{V} \mathbf{T}^{\text{T}} + \mathbf{N}.\]

Example

julia> noise = [1.0 -3.0; 4.0 2.0];

julia> displace(1.0+im, noise)
GaussianChannel for 1 mode.
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
transform: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0
noise: 2×2 Matrix{Float64}:
 1.0  -3.0
 4.0   2.0

Listed below are a list of predefined Gaussian channels supported by Gabs.jl:

attenuator

amplifier

Note

When its noise matrix $\mathbf{N} = \mathbf{0}$ and transform operator $\mathbf{T}$ is a symplectic matrix, a Gaussian channel is a unitary operator. Any predefined Gaussian unitary method can be called with an additional noise matrix to create a GaussianChannel object. For instance, a noisy displacement operator can be called with displace as follows:

julia> noise = [1.0 -2.0; 4.0 -3.0];

julia> displace(1.0-im, noise)
GaussianChannel for 1 mode.
displacement: 2-element Vector{Float64}:
  1.4142135623730951
 -1.4142135623730951
transform: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0
noise: 2×2 Matrix{Float64}:
 1.0  -2.0
 4.0  -3.0