Gaussian Zoos

State Zoo

Gabs.vacuumstate — Function

vacuumstate([Tm=Vector{Float64}, Tc=Matrix{Float64}], basis::SymplecticBasis)

Gaussian state with zero photons, known as the vacuum state. The symplectic representation is defined by basis.

Mathematical description of a vacuum state

A vacuum state $|0\rangle$ is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> vacuumstate(QuadPairBasis(1))
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.thermalstate — Function

thermalstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, photons<:Int)

Gaussian state at thermal equilibrium, known as the thermal state. The symplectic representation is defined by basis. The mean photon number of the state is given by photons.

Mathematical description of a thermal state

A thermal state $|\bar{n}\rangle$, where $\bar{n}$ is the mean number of photons, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \left(\bar{n} + \frac{1}{2}\right)\mathbf{I}.\]

Example

julia> thermalstate(QuadPairBasis(1), 4)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 4.5  0.0
 0.0  4.5

Gabs.coherentstate — Function

coherentstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, alpha<:Number)

Gaussian state that is the quantum analogue of a monochromatic electromagnetic field, known as the coherent state. The symplectic representation is defined by basis. The complex amplitude of the state is given by alpha.

Mathematical description of a coherent state

A coherent state $|\alpha\rangle$, where $\alpha$ is the complex amplitude, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{V} = \mathbf{I}.\]

Example

julia> coherentstate(QuadPairBasis(1), 1.0+im)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
covariance: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.squeezedstate — Function

squeezedstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, r<:Real, theta<:Real)

Gaussian state with quantum uncertainty in its phase and amplitude, known as the squeezed state. The symplectic representation is defined by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of a squeezed state

A squeezed state $|r, \theta\rangle$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\left(\cosh(2r)\mathbf{I} - \sinh(2r)\mathbf{R}(\theta)\right),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeezedstate(QuadPairBasis(1), 0.5, pi/4)
GaussianState for 1 mode.
  symplectic basis: QuadPairBasis
mean: 2-element Vector{Float64}:
 0.0
 0.0
covariance: 2×2 Matrix{Float64}:
 0.356044  0.415496
 0.415496  1.18704

Gabs.eprstate — Function

eprstate([Tm=Vector{Float64}, Tc=Matrix{Float64},] basis::SymplecticBasis, r<:Real, theta<:Real)

Gaussian state that is a two-mode squeezed state, known as the Einstein-Podolski-Rosen (EPR) state. The symplectic representation is defined by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively.

Mathematical description of an EPR state

An EPR state $|r, \theta\rangle_{\text{EPR}}$, where $r$ is the amplitude squeezing parameter and $\theta$ is the phase squeezing parameter, is characterized by the mean vector $\mathbf{\bar{x}}$ and covariance matrix $\mathbf{V}$, expressed respectively as follows:

\[\mathbf{\bar{x}} = \mathbf{0}, \quad \mathbf{V} = \frac{1}{2}\begin{pmatrix} \cosh(2r)\mathbf{I} & -\sinh(2r)\mathbf{R}(\theta) \\ -\sinh(2r)\mathbf{R}(\theta) & \cosh(2r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> eprstate(QuadPairBasis(2), 0.5, pi/4)
GaussianState for 2 modes.
  symplectic basis: QuadPairBasis
mean: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
covariance: 4×4 Matrix{Float64}:
  0.77154    0.0       -0.415496  -0.415496
  0.0        0.77154   -0.415496   0.415496
 -0.415496  -0.415496   0.77154    0.0
 -0.415496   0.415496   0.0        0.77154

Operator Zoo

Gabs.displace — Function

displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, alpha<:Number)
displace([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, alpha<:Number, noise::Ts)

Gaussian operator that displaces the vacuum state into a coherent state, known as the displacement operator. The symplectic representation is given by basis. The complex amplitude is given by alpha. Noise can be added to the operation with noise.

Mathematical description of a displacement operator

A displacement operator $D(\alpha)$ is defined by the operation $D(\alpha)|0\rangle = |\alpha\rangle$, where $\alpha$ is a complex amplitude. The operator $D(\alpha)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \sqrt{2}\left(\text{Re}(\alpha), \text{Im}(\alpha)\right)^{\text{T}}, \quad \mathbf{S} = \mathbf{I}.\]

Example

julia> displace(QuadPairBasis(1), 1.0+im)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 1.4142135623730951
 1.4142135623730951
symplectic: 2×2 Matrix{Float64}:
 1.0  0.0
 0.0  1.0

Gabs.squeeze — Function

squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real)
squeeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes the vacuum state into a squeezed state, known as the squeezing operator. The symplectic representation is given by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a squeezing operator

A squeeze operator $S(r, \theta)$ is defined by the operation $S(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \cosh(r)\mathbf{I} - \sinh(r)\mathbf{R}(\theta),\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> squeeze(QuadPairBasis(1), 0.25, pi/4)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
  0.852789  -0.178624
 -0.178624   1.21004

Gabs.twosqueeze — Function

twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real)
twosqueeze([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, r<:Real, theta<:Real, noise::Ts)

Gaussian operator that squeezes a two-mode vacuum state into a two-mode squeezed state, known as the two-mode squeezing operator. The symplectic representation is given by basis. The amplitude and phase squeezing parameters are given by r and theta, respectively. Noise can be added to the operation with noise.

Mathematical description of a two-mode squeezing operator

A two-mode squeeze operator $S_2(r, \theta)$ is defined by the operation $S_2(r, \theta)|0\rangle = |r, \theta\rangle$, where $r$ and $\theta$ are the real amplitude and phase parameters, respectively. The operator $S_2(r, \theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cosh(r)\mathbf{I} & -\sinh(r)\mathbf{R}(\theta) \\ -\sinh(r)\mathbf{R}(\theta) & \cosh(r)\mathbf{I} \\ \end{pmatrix},\]

where $\mathbf{R}(\theta)$ is the rotation matrix.

Example

julia> twosqueeze(QuadPairBasis(2), 0.25, pi/4)
GaussianUnitary for 2 modes.
  symplectic basis: QuadPairBasis
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  1.03141    0.0       -0.178624  -0.178624
  0.0        1.03141   -0.178624   0.178624
 -0.178624  -0.178624   1.03141    0.0
 -0.178624   0.178624   0.0        1.03141

Gabs.phaseshift — Function

phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, theta<:Real)
phaseshift([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, theta<:Real, noise::Ts)

Gaussian operator that rotates the phase of a given Gaussian mode by theta, as the phase shift operator. The symplectic representation is given by basis. Noise can be added to the operation with noise.

Mathematical description of a phase shift operator

A phase shift operator $U(\theta)$ is defined by the operation $U(\Theta) = \exp(-i\theta\hat{a}^{\dagger}\hat{a})$, where $\theta$ is the phase parameter, and $\hat{a}^{\dagger}$ and $\hat{a}$ are the raising and lowering operators, respectively. The operator $U(\theta)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \cos(\theta) & \sin(\theta) \\ -\sin(\theta) & \cos(\theta) \\ \end{pmatrix}.\]

Example

julia> phaseshift(QuadPairBasis(1), 3pi/4)
GaussianUnitary for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
symplectic: 2×2 Matrix{Float64}:
 -0.707107   0.707107
 -0.707107  -0.707107

Gabs.beamsplitter — Function

beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, transmit<:Real)
beamsplitter([Tm=Vector{Float64}, Ts=Matrix{Float64}], basis::SymplecticBasis, transmit<:Real, noise::Ts)

Gaussian operator that serves as the beam splitter transformation of a two-mode Gaussian state, known as the beam splitter operator. The symplectic representation is given by basis. The transmittivity of the operator is given by transmit. Noise can be added to the operation with noise.

Mathematical description of a beam splitter operator

A beam splitter operator $B(\tau)$ is defined by the operation $B(\theta) = \exp\left[\theta(\hat{a}^{\dagger}\hat{b} - \hat{a}\hat{b}^{\dagger})\right]$, where $\theta$ is defined by $\tau = \cos^2\theta$, and $\hat{a}$ and $\hat{b}$ are the annihilation operators of the two modes, respectively. The operator $B(\tau)$ is characterized by the displacement vector $\mathbf{d}$ and symplectic matrix $\mathbf{S}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{S} = \begin{pmatrix} \sqrt{\tau}\mathbf{I} & \sqrt{1-\tau}\mathbf{I} \\ -\sqrt{1-\tau}\mathbf{I} & \sqrt{\tau}\mathbf{I} \\ \end{pmatrix}.\]

Example

julia> beamsplitter(QuadPairBasis(2), 0.75)
GaussianUnitary for 2 modes.
  symplectic basis: QuadPairBasis
displacement: 4-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
symplectic: 4×4 Matrix{Float64}:
  0.5        0.0       0.866025  0.0
  0.0        0.5       0.0       0.866025
 -0.866025   0.0       0.5       0.0
  0.0       -0.866025  0.0       0.5

Gabs.attenuator — Function

attenuator([Td=Vector{Float64}, Tt=Matrix{Float64},] basis::SymplecticBasis, theta<:Real, n<:Int)

Gaussian channel describing the coupling of an input single mode Gaussian state and its environment via a beam splitter operation. The channel is paramatrized by beam splitter rotation angle theta and thermal noise n.

Mathematical description of an attenuator channel

An attenuator channel, $\mathcal{E}_{\theta}^{n_{\text{th}}}$, where $\theta$ is the beam splitter rotation parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cos\theta\mathbf{I}, \qquad \mathbf{N} = (\sin\theta)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> attenuator(QuadPairBasis(1), pi/6, 3)
GaussianChannel for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 0.866025  0.0
 0.0       0.866025
noise: 2×2 Matrix{Float64}:
 0.75  0.0
 0.0   0.75

Gabs.amplifier — Function

amplifier([Td=Vector{Float64}, Tt=Matrix{Float64},] basis::SymplecticBasis, r<:Real, n<:Int)

Gaussian channel describing the interaction of an input single mode Gaussian state and its environment via a two-mode squeezing operation. The channel is paramatrized by squeezing amplitude parameter r and thermal noise n.

Mathematical description of an amplifier channel

An amplifier channel, $\mathcal{A}_{\theta}^{n_{\text{th}}}$, where $r$ is the squeezing amplitude parameter and $n_{\text{th}} \geq 1$ is the thermal noise parameter, is characterized by the displacement vector $\mathbf{d}$, transformation matrix $\mathbf{T}$, and noise matrix $\mathbf{N}$, expressed respectively as follows:

\[\mathbf{d} = \mathbf{0}, \quad \mathbf{T} = \cosh r\mathbf{I}, \qquad \mathbf{N} = (\sinh r)^2 n_{\text{th}} \mathbf{I}.\]

Example

julia> amplifier(QuadPairBasis(1), 2.0, 3)
GaussianChannel for 1 mode.
  symplectic basis: QuadPairBasis
displacement: 2-element Vector{Float64}:
 0.0
 0.0
transform: 2×2 Matrix{Float64}:
 3.7622  0.0
 0.0     3.7622
noise: 2×2 Matrix{Float64}:
 39.4623   0.0
  0.0     39.4623